+++ src/types.h
@@ -327,7 +327,7 @@ typedef int16_t        PieceToHistory[16][64];
 typedef PieceToHistory CounterMoveHistoryStat[16][64];
 typedef int16_t        ButterflyHistory[2][4096];
 typedef int16_t        CapturePieceToHistory[16][64][8];
-typedef int32_t        correction_history_t[2][CORRECTION_HISTORY_ENTRY_NB];
+typedef int16_t        correction_history_t[2][CORRECTION_HISTORY_ENTRY_NB];
 
 struct ExtMove {
     Move move;
